chore(search-api): Move files to packages according to types
