+++ applications/search-api/src/main/java/no/fdk/searchapi/controller/CatalogController.java
@@ -2,8 +2,6 @@ package no.fdk.searchapi.controller;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import no.fdk.searchapi.ModelFormatter;
-import no.fdk.searchapi.SupportedFormat;
 import no.dcat.shared.Catalog;
 import no.dcat.shared.Dataset;
 import org.apache.jena.query.*;
