+++ applications/search-api/src/main/java/no/fdk/searchapi/controller/DatasetsSearchController.java
@@ -2,7 +2,7 @@ package no.fdk.searchapi.controller;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import no.fdk.searchapi.ElasticsearchService;
+import no.fdk.searchapi.service.ElasticsearchService;
 import no.dcat.shared.Dataset;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
