+++ applications/search-api/src/test/java/no/fdk/searchapi/ModelFormatterTest.java
@@ -1,5 +1,7 @@
 package no.fdk.searchapi;
 
+import no.fdk.searchapi.controller.ModelFormatter;
+import no.fdk.searchapi.controller.SupportedFormat;
 import no.fdk.test.testcategories.UnitTest;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
